refactor Parallel
