+++ src/Approximations/Approximations.jl
@@ -12,13 +12,11 @@ export approximate,
        ballinf_approximation,
        box_approximation,
        decompose,
-       decompose_parallel,
        diameter,
        norm,
        overapproximate,
        radius,
        symmetric_interval_hull,
-       symmetric_interval_hull_parallel,
        BoxDirections,
        BoxDiagDirections,
        OctDirections
